+++ tests/test_functional/test_retry_if_errors.py
@@ -12,7 +12,7 @@ def test_catch_error_with_tries():
 
     with pytest.raises(ValueError):
         decorated_func()
-    assert func.call_count == 4
+    assert func.call_count == 3
 
 
 def test_catch_many_errors_with_tries():
@@ -21,7 +21,7 @@ def test_catch_many_errors_with_tries():
 
     with pytest.raises(ValueError):
         decorated_func()
-    assert func.call_count == 4
+    assert func.call_count == 3
 
 
 def test_catch_error_without_tries():
@@ -30,7 +30,7 @@ def test_catch_error_without_tries():
 
     with pytest.raises(ValueError):
         decorated_func()
-    assert func.call_count == 2
+    assert func.call_count == 1
 
 
 def test_not_raise_error():
@@ -57,3 +57,11 @@ def test_not_catch_error():
     with pytest.raises(ValueError):
         decorated_func()
     assert func.call_count == 1
+
+
+@pytest.mark.parametrize("tries", [0, -1])
+def test_raises_error_on_negative_or_zero_tries(tries):
+    """Test that the function raises a ValueError when the number of tries is zero or negative."""
+
+    with pytest.raises(ValueError):
+        pymince.functional.retry_if_errors(tries=tries)