+++ tests/test_json/test_csv_to_json.py
@@ -39,7 +39,7 @@ def test_dumped_data(extension):
         {"co l3": "%6$", "col1": "baz", "col2": "foo", "col4": "567"},
     ]
     with temp_paths(extension) as (csv_path, json_path):
-        pymince.json.dump_from_csv(csv_path, json_path, fieldnames=csv_head)
+        pymince.json.csv_to_json(csv_path, json_path, fieldnames=csv_head)
         dumped = pymince.json.load_from(json_path)
         assert dumped == expected
 
@@ -50,7 +50,7 @@ def test_dumped_data_excluding_header():
         {"co l3": "%6$", "col1": "baz", "col2": "foo", "col4": "567"},
     ]
     with temp_paths(".json") as (csv_path, json_path):
-        pymince.json.dump_from_csv(csv_path, json_path, fieldnames=csv_head, start=1)
+        pymince.json.csv_to_json(csv_path, json_path, fieldnames=csv_head, start=1)
         dumped = pymince.json.load_from(json_path)
         assert dumped == expected
 
@@ -60,7 +60,7 @@ def test_dumped_data_excluding_header_and_footer():
         {"co l3": "!!!", "col1": "ñoo", "col2": "var", "col4": "123"},
     ]
     with temp_paths(".json") as (csv_path, json_path):
-        pymince.json.dump_from_csv(csv_path, json_path, fieldnames=csv_head, start=1, stop=2)
+        pymince.json.csv_to_json(csv_path, json_path, fieldnames=csv_head, start=1, stop=2)
         dumped = pymince.json.load_from(json_path)
         assert dumped == expected
 
@@ -68,7 +68,7 @@ def test_dumped_data_excluding_header_and_footer():
 def test_dumped_data_start_and_stop_equals():
     expected = []
     with temp_paths(".json") as (csv_path, json_path):
-        pymince.json.dump_from_csv(csv_path, json_path, fieldnames=csv_head, start=1, stop=1)
+        pymince.json.csv_to_json(csv_path, json_path, fieldnames=csv_head, start=1, stop=1)
         dumped = pymince.json.load_from(json_path)
         assert dumped == expected
 
@@ -80,7 +80,7 @@ def test_dumped_data_strip_false():
         {" col1": " baz", " col4 ": "567", "co l3": " %6$ ", "col2 ": "foo "},
     ]
     with temp_paths(".json") as (csv_path, json_path):
-        pymince.json.dump_from_csv(csv_path, json_path, fieldnames=csv_head, strip=False)
+        pymince.json.csv_to_json(csv_path, json_path, fieldnames=csv_head, strip=False)
         dumped = pymince.json.load_from(json_path)
         assert dumped == expected
 
@@ -93,7 +93,7 @@ def test_dumped_data_with_fieldnames():
         {"key1": "baz", "key2": "foo", "key3": "%6$", "key4": "567"},
     ]
     with temp_paths(".json") as (csv_path, json_path):
-        pymince.json.dump_from_csv(csv_path, json_path, fieldnames=fieldnames)
+        pymince.json.csv_to_json(csv_path, json_path, fieldnames=fieldnames)
         dumped = pymince.json.load_from(json_path)
         assert dumped == expected
 
@@ -106,7 +106,7 @@ def test_dumped_data_with_null_keys():
         {"key1": "baz", "key2": "foo", "null": [" %6$ ", "567"]},
     ]
     with temp_paths(".json") as (csv_path, json_path):
-        pymince.json.dump_from_csv(csv_path, json_path, fieldnames=fieldnames)
+        pymince.json.csv_to_json(csv_path, json_path, fieldnames=fieldnames)
         dumped = pymince.json.load_from(json_path)
         assert dumped == expected
 
@@ -119,7 +119,7 @@ def test_dumped_data_with_null_values():
         {"key1": "baz", "key2": "foo", "key3": "%6$", "key4": "567", "key5": None},
     ]
     with temp_paths(".json") as (csv_path, json_path):
-        pymince.json.dump_from_csv(csv_path, json_path, fieldnames=fieldnames)
+        pymince.json.csv_to_json(csv_path, json_path, fieldnames=fieldnames)
         dumped = pymince.json.load_from(json_path)
         assert dumped == expected
 
@@ -131,6 +131,6 @@ def test_dumped_data_with_empty_fields():
         {"null": [" baz", "foo ", " %6$ ", "567"]},
     ]
     with temp_paths(".json") as (csv_path, json_path):
-        pymince.json.dump_from_csv(csv_path, json_path, fieldnames=[])
+        pymince.json.csv_to_json(csv_path, json_path, fieldnames=[])
         dumped = pymince.json.load_from(json_path)
         assert dumped == expected